pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.poetry]
authors = ["Alexander Belikov <abelikov@gmail.com>"]
description = ""
name = "xsi"
packages = [{include = "xsi"}]
readme = "README.md"
version = "0.0.1"

[tool.poetry.dependencies]
Flask = "^3.1.0"
Flask-Cors = "^5.0.0"
Flask-RESTful = "^0.3.10"
Sickle = "^0.7.0"
aiohttp = "^3.11.9"
aiohttp-retry = "^2.9.1"
aiolimiter = "^1.2.0"
arxiv = "^2.1.3"
beautifulsoup4 = "^4.12.2"
bidict = "^0.23.1"
biopython = "^1.84"
certifi = "^2024.8.30"
click = "^8.1.7"
curlify = "^2.2.1"
dataclass-wizard = {extras = ["timedelta"], version = "0.34.0"}
dill = "^0.3.7"
flask-compress = "^1.13"
flask-limiter = "^3.8.0"
graph-cast = {git = "ssh://git@github.com/growgraph/graph_cast.git", rev = "v0.13.19"}
igraph = "^0.11.6"
isoweek = "^1.3.3"
nltk = "^3.7"
numpy = "^2.1.3"
ortools = "^9.9.3963"
pandas = "^2.2.3"
pathos = "^0.3.1"
pip = "^25.0"
polars = "^1.11.0"
prometheus-flask-exporter = "^0.23.1"
pyarrow = "^17.0.0"
pyhumps = "^3.8.0"
pympler = "^1.0.1"
pysbd = "^0.3.4"
python = ">=3.11,<3.12"
requests = "^2.27.1"
scikit-learn = "1.5.2"
scipy = "^1.12.0"
setuptools = "^75.2.0"
statsmodels = "^0.14.4"
tabulate = "^0.9.0"
torch = [
  {markers = "platform_machine == 'aarch64' and sys_platform != 'darwin'", source = "pypi", version = "^2.4.0"},
  {markers = "platform_machine == 'x86_64' and sys_platform != 'darwin'", source = "pytorch_cpu", version = "^2.4.0"},
  {markers = "sys_platform == 'darwin'", source = "pypi", version = "^2.4.0"}
]
torchvision = [
  {markers = "platform_machine == 'aarch64' and sys_platform != 'darwin'", source = "pypi", version = "^0.19.0"},
  {markers = "platform_machine == 'x86_64' and sys_platform != 'darwin'", source = "pytorch_cpu", version = "^0.19.0"},
  {markers = "sys_platform == 'darwin'", source = "pypi", version = "^0.19.0"}
]
waitress = "^3.0.0"
wheel = "^0.44.0"

[tool.poetry.group.dev.dependencies]
git-fame = "^2.0.2"
kaleido = "0.2.1"
matplotlib = "^3.9.2"
plotly = "^5.24.1"
pre-commit = "^3.3.3"
pygraphviz = "^1.13"
pylint = "^3.2.6"
pytest = "^8.3.3"
seaborn = "^0.13.2"

[[tool.poetry.source]]
name = "pytorch_cpu"
priority = "explicit"
url = "https://download.pytorch.org/whl/cpu"
